Avoid picking a non-existing root if one existing on the same depth
If mail A is answer to B and C, and C does not exist, always pick B.
Non-existing roots are only accepted if there is no other mail in the thread
above or next to it.

Corner case: Existing thread root has more than one non-existing ancestor.
Then, the result is deterministic (due to “root_ > root”), but may be
incorrect (because I don’t look at the In-Reply-To field).

This way, I can guarantee that all mails in a thread point to the same root,
which was not assured before.