Fix thread root finding for good
In the end, only a brute force algorithm worked.  I now collect the whole
thread by flood-walking through it, and then I find the node (aka mail) with
the biggest sub-thread.  For the latter, I have to prefer existing mails, and
if only non-existing have the biggest set of descendants, if take the one with
the lexicographically first hash ID.  All of this is necessary to avoid
inconsistent or unstable root-detection.